@@ -2,6 +2,6 @@
 name = "optlt"
 version = "0.1.0"
 authors = ["French Bear <FrenchBear38@outlook.com>"]
-edition = "2021"
+edition = "2024"
 
 [dependencies]
